Ratelimit measurements from sensors
Previously on the Arduino, there was no limit to how fast the system
could query the sensors so tests in the lab did not reflect real world
examples. By adding a rate limit, this insures the system will act
in the same manner wherever it is located.
This commit also changes all time to use unsigned longs, relecting
Arduino standards.